// app/src/generated/models/Al_signoffsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Al_signoffsal_signoffdecision = {
  120910720: 'Approved',
  120910721: 'Rejected'
} as const;
export type Al_signoffsal_signoffdecision = keyof typeof Al_signoffsal_signoffdecision;
export const Al_signoffsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Al_signoffsstatecode = keyof typeof Al_signoffsstatecode;
export const Al_signoffsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Al_signoffsstatuscode = keyof typeof Al_signoffsstatuscode;

export interface Al_signoffsBase {
  al_name: string;
  al_notes?: string;
  "al_outcomecaseid@odata.bind": string;
  "al_remediationactionid@odata.bind"?: string;
  al_signedoffon?: string;
  al_signoffcode: string;
  al_signoffdecision: Al_signoffsal_signoffdecision;
  al_signoffid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Al_signoffsstatecode;
  statuscode?: Al_signoffsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Al_signoffs extends Al_signoffsBase {
  al_outcomecaseidname?: string;
  al_remediationactionidname?: string;
  al_signoffdecisionname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  al_outcomecaseid?: object;
  _al_outcomecaseid_value?: string;
  al_remediationactionid?: object;
  _al_remediationactionid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}